requirements.txt
pyquaternion
pyyaml
opencv-python-headless
matplotlib==3.5.2
nuscenes-devkit==1.1.10
tensorboard==2.12.0
tensorboardX==2.6
numpy==1.19.0
scipy==1.9.3
pandas==1.4.4
